requirements.txt
# Flask 核心
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# 数据库
SQLAlchemy==2.0.23
alembic==1.13.0

# 数据验证与序列化
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# 密码加密
bcrypt==4.1.2

# 数据处理
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# 机器学习
scikit-learn==1.3.2
tensorflow==2.15.0
numpy==1.26.2

# 可视化与报告
matplotlib==3.8.2
reportlab==4.0.7
WeasyPrint==60.1
python-docx==1.1.0
Pillow==10.1.0

# 工具库
python-dotenv==1.0.0
gunicorn==21.2.0

# 测试
pytest==7.4.3
pytest-cov==4.1.0

